+++ tests/Checks/IsEventNameTest.php
@@ -1,0 +1,190 @@
+<?php
+
+/**
+ * Copyright (c) 2015-present Ganbaro Digital Ltd
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions
+ * are met:
+ *
+ *   * Redistributions of source code must retain the above copyright
+ *     notice, this list of conditions and the following disclaimer.
+ *
+ *   * Redistributions in binary form must reproduce the above copyright
+ *     notice, this list of conditions and the following disclaimer in
+ *     the documentation and/or other materials provided with the
+ *     distribution.
+ *
+ *   * Neither the names of the copyright holders nor the names of his
+ *     contributors may be used to endorse or promote products derived
+ *     from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
+ * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
+ * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS
+ * FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE
+ * COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT,
+ * INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
+ * BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER
+ * CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT
+ * LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
+ * ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ *
+ * @category  Libraries
+ * @package   EventStream/Streams
+ * @author    Stuart Herbert <stuherbert@ganbarodigital.com>
+ * @copyright 2015-present Ganbaro Digital Ltd www.ganbarodigital.com
+ * @license   http://www.opensource.org/licenses/bsd-license.php  BSD License
+ * @link      http://code.ganbarodigital.com/php-event-stream
+ */
+
+namespace GanbaroDigital\EventStream\Checks;
+
+use GanbaroDigital\EventStream\Checks\HasEventHandler;
+use GanbaroDigital\EventStream\Events\Event;
+use PHPUnit_Framework_TestCase;
+use RuntimeException;
+
+class IsEventNameTest_DummyEvent implements Event { }
+
+/**
+ * @coversDefaultClass GanbaroDigital\EventStream\Checks\IsEventName
+ */
+class IsEventNameTest extends PHPUnit_Framework_TestCase
+{
+    /**
+     * @coversNothing
+     */
+    public function testCanInstantiate()
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $obj = new IsEventName();
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($obj instanceof IsEventName);
+    }
+
+    /**
+     * @covers ::__invoke
+     * @dataProvider provideDataToTest
+     */
+    public function testCanUseAsObject($eventName, $expectedResult)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        $obj = new IsEventName();
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = $obj($eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $actualResult);
+    }
+
+
+    /**
+     * @covers ::check
+     * @dataProvider provideDataToTest
+     */
+    public function testCanCallStatically($eventName, $expectedResult)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = IsEventName::check($eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertEquals($expectedResult, $actualResult);
+    }
+
+    /**
+     * @covers ::check
+     * @dataProvider provideValidEventNames
+     */
+    public function testReturnsTrueForValidEventNames($eventName, $expectedResult)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = IsEventName::check($eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertTrue($actualResult);
+    }
+
+    /**
+     * @covers ::check
+     * @dataProvider provideInvalidEventNames
+     */
+    public function testReturnsFalseForInvalidEventNames($eventName, $expectedResult)
+    {
+        // ----------------------------------------------------------------
+        // setup your test
+
+        // ----------------------------------------------------------------
+        // perform the change
+
+        $actualResult = IsEventName::check($eventName);
+
+        // ----------------------------------------------------------------
+        // test the results
+
+        $this->assertFalse($actualResult);
+    }
+
+    public function provideDataToTest()
+    {
+        return [
+            [ null, false ],
+            [ [], false ],
+            [ true, false ],
+            [ false, false ],
+            [ function() {}, false ],
+            [ 3.1415927, false ],
+            [ 0, false ],
+            [ 100, false ],
+            [ new IsEventName, false ],
+            [ "hello, world!", false ],
+            [ IsEventNameTest_DummyEvent::class, true ],
+        ];
+    }
+
+    public function provideValidEventNames()
+    {
+        return [
+            [ IsEventNameTest_DummyEvent::class, true ],
+        ];
+    }
+
+    public function provideInvalidEventNames()
+    {
+        return [
+            [ IsEventName::class, false ],
+        ];
+    }
+}