+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### New
 
 * Checks\HasEventHandler
+* Checks\IsEventName
 * Events\Event
 * Exceptions\Exxx_EventStreamException
 * Exceptions\E4xx_EventStreamException